+++ tfs/handler.py
@@ -91,10 +91,20 @@ class TfsDataFrame(pd.DataFrame):
             s += "\n"
         return f"{s}{super().__repr__()}"
 
+    def check_unique_indices(self):
+        """Checks that the object has unique indices, raises a ``TfsFormatError`` if it does not."""
+        if self.index.has_duplicates:
+            LOGGER.error("Non-unique indices found")
+            raise TfsFormatError("Indices are not unique.")
 
-def read_tfs(
-    tfs_file_path: Union[pathlib.Path, str], index: str = None, check_unique_indices: bool = False
-) -> TfsDataFrame:
+    def check_unique_columns(self):
+        """Checks that the object has unique columns, raises a ``TfsFormatError`` if it does not."""
+        if self.columns.has_duplicates:
+            LOGGER.error("Non-unique columns found")
+            raise TfsFormatError("Columns are not unique")
+
+
+def read_tfs(tfs_file_path: Union[pathlib.Path, str], index: str = None) -> TfsDataFrame:
     """
     Parses the TFS table present in **tfs_file_path** and returns a customized version of a Pandas
     DataFrame (a TfsDataFrame).
@@ -104,8 +114,6 @@ def read_tfs(
             a string, in which case it will be cast to a PosixPath object.
         index (str): Name of the column to set as index. If not given, looks in **tfs_file_path**
             for a column starting with `INDEX&&&`.
-        check_unique_indices (bool): Flag to check that the read file has unique indices. If set to ``True``
-            and indices are not unique, a ``TfsFormatError`` will be raised. Defaults to False.
 
     Returns:
         A TfsDataFrame object.
@@ -154,7 +162,7 @@ def read_tfs(
                 index_name = None  # to remove it completely (Pandas makes a difference)
             data_frame = data_frame.rename_axis(index_name)
 
-    _validate(data_frame, f"from file {tfs_file_path.absolute()}", check_unique_indices)
+    _validate(data_frame, f"from file {tfs_file_path.absolute()}")
     return data_frame
 
 
@@ -165,7 +173,6 @@ def write_tfs(
     save_index: Union[str, bool] = False,
     colwidth: int = DEFAULT_COLUMN_WIDTH,
     headerswidth: int = DEFAULT_COLUMN_WIDTH,
-    check_unique_indices: bool = False,
 ) -> None:
     """
     Writes the DataFrame into **tfs_file_path** with the `headers_dict` as headers dictionary. If
@@ -183,13 +190,10 @@ def write_tfs(
             saves the index of the data_frame to a column named by the provided value.
         colwidth (int): Column width, can not be smaller than `MIN_COLUMN_WIDTH`.
         headerswidth (int): Used to format the header width for both keys and values.
-        check_unique_indices (bool): Flag to check that the dataframe to be written has unique indices. If
-            set to ``True`` and indices are not unique, a ``TfsFormatError`` will be raised. Defaults to
-            ``False``.
     """
     left_align_first_column = False
     tfs_file_path = pathlib.Path(tfs_file_path)
-    _validate(data_frame, f"to be written in {tfs_file_path.absolute()}", check_unique_indices)
+    _validate(data_frame, f"to be written in {tfs_file_path.absolute()}")
 
     if headers_dict is None:  # tries to get headers from TfsDataFrame
         try:
@@ -446,16 +450,14 @@ class TfsFormatError(Exception):
 def _validate(
     data_frame: Union[TfsDataFrame, pd.DataFrame],
     info_str: str = "",
-    check_unique_indices: bool = False,
 ) -> None:
     """
-    Check if Dataframe contains finite values only and both indices and columns are unique.
+    Check if Dataframe contains finite values only, strings as column names and no empty headers
+    or column names.
 
     Args:
         data_frame (Union[TfsDataFrame, pd.DataFrame]): the dataframe to check on.
         info_str (str): additional information to includ in logging statements.
-        check_unique_indices (bool): Flag to check that the dataframe has unique indices. If set to
-            ``True`` and indices are not unique, a ``TfsFormatError`` will be raised. Defaults to ``False``.
     """
 
     def is_not_finite(x):
@@ -474,14 +476,6 @@ def _validate(
             f"DataFrame {info_str} contains non-physical values at Index: "
             f"{boolean_df.index[boolean_df.any(axis='columns')].tolist()}"
         )
-
-    if check_unique_indices and len(set(data_frame.index)) != len(data_frame.index):
-        LOGGER.error(f"Non-unique indices found, dataframe {info_str} is invalid")
-        raise TfsFormatError("Indices are not Unique.")
-
-    if len(set(data_frame.columns)) != len(data_frame.columns):
-        LOGGER.error(f"Non-unique column names found, dataframe {info_str} is invalid")
-        raise TfsFormatError("Column names not Unique.")
 
     if any(not isinstance(c, str) for c in data_frame.columns):
         LOGGER.error(f"Some column-names are not of string-type, dataframe {info_str} is invalid.")